adding a priceService
